Solve day 16
I had actually solved this before, though I don't remember when or in
what language. I think I did it because it was one of these
instruction-execution problems which I quite like.

Getting the process-of-elimination solve just right took a lot of
fiddling. The code is not pretty but it works. If a particular opcode
only has one candidate operation, then that must be it. Likewise, if a
particular operation is only the candidate for one operation, then that
must be it.