Day 21 part 2
I like when a part 2 is simple.

We really only have two variables (three and four), and four is
redefined at the top of each outer loop. That means that the entirety of
the program state is captured in register three. So if register three
attains a value that it's previously attained, then we've entered an
endless loop.

If we let the program enter an endless loop, it never terminates. The
latest possible termination is right before entering this loop. We want
to break when three attains whatever value it attains just before
revisiting a previous state.