Day 22 part 2, finally!
I read some solutions online and they all used a priority queue. I
thought this approach should still be viable.

With fresh eyes, I managed to figure it out. There's tricky conditions
around managing "next_gen" (which is the set of points to process in the
next round). We don't want to accidentally override a shorter route to a
particular square.

I think I was dancing around this for a while and trying several
slightly different approaches, but finally the pieces came together.