Day 24 part 2
We can just do a linear search.

This completes in under 2 seconds. We can save a second by using a
binary search, but that's more code for not that much more performance.

I had to handle the "stalemate" case, so that we don't end up in an
infinite loop. This wasn't needed in part 1.